Extract HTTP/TLS checks into NetworkUtils
