+++ src/Certify.Core/Management/NetworkUtils.cs
@@ -1,0 +1,166 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Net;
+using System.Net.Http;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Certify.Management
+{
+    public class NetworkUtils
+    {
+        public Action<string> Log = (message) => { };
+
+        public bool CheckSNI(string host, string sni, bool? useProxyAPI = null)
+        {
+            // if validation proxy enabled, access to the domain being validated is checked via our
+            // remote API rather than directly on the servers
+            bool useProxy = useProxyAPI ?? Certify.Properties.Settings.Default.EnableValidationProxyAPI;
+            if (useProxy)
+            {
+                // TODO: check proxy here, needs server support. if successful "return true"; and "LogAction(...)"
+                System.Diagnostics.Debug.WriteLine("ProxyAPI is not implemented for Checking SNI config, trying local");
+                Log($"Proxy TLS SNI binding check error: {host}, {sni}");
+
+                return CheckSNI(host, sni, false); // proxy failed, try local
+            }
+            var hosts = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.System), @"drivers\etc\hosts");
+            try
+            {
+                var req = new HttpRequestMessage(HttpMethod.Get, $"https://{sni}");
+                ServicePointManager.ServerCertificateValidationCallback = (obj, cert, chain, errors) =>
+                {
+                    // verify SNI-selected certificate is correctly configured
+                    return CertificateManager.VerifyCertificateSAN(cert, sni);
+                };
+
+                // modify the hosts file so we can resolve this request locally: create an entry for
+                // the primary IP address and also for 127.0.0.1 (where primary IP will not resolve
+                // internally i.e. the default resolution is an external IP)
+
+                List<string> testHostEntries = new List<string> {
+                    $"\n127.0.0.1\t{sni}",
+                };
+
+                var ip = Dns.GetHostEntry(host)?.AddressList?.FirstOrDefault();
+                if (ip != null)
+                {
+                    testHostEntries.Add($"\n{ip}\t{sni}");
+                }
+
+                using (StreamWriter writer = File.AppendText(hosts))
+                {
+                    foreach (var hostEntry in testHostEntries)
+                    {
+                        writer.Write(hostEntry);
+                    }
+                }
+                Thread.Sleep(250); // wait a bit for hosts file to take effect
+
+                try
+                {
+                    using (var client = new HttpClient())
+                    {
+                        var resp = client.SendAsync(req).Result;
+                        // if the GET request succeeded, the Cert validation succeeded
+                        Log($"Local TLS SNI binding check OK: {host}, {sni}"); ;
+                    }
+                }
+                finally
+                {
+                    // clean up temp entries in hosts file
+                    try
+                    {
+                        var txt = File.ReadAllText(hosts);
+                        foreach (var hostEntry in testHostEntries)
+                        {
+                            //should we just remove all .acme.invalid entries instead of looking for our current entries?
+                            txt = txt.Substring(0, txt.Length - hostEntry.Length);
+                        }
+
+                        File.WriteAllText(hosts, txt);
+                    }
+                    catch
+                    {
+                        // if this fails the user will have to clean up manually
+                        Log($"Error cleaning up hosts file: {hosts}");
+                        throw;
+                    }
+                }
+                return true; // success!
+            }
+            catch (Exception ex)
+            {
+                // eat the error that HttpClient throws, either cert validation failed or the site is
+                // inaccessible via https://host name
+                Log($"Local TLS SNI binding check error: {host}, {sni}\n{ex.GetType()}: {ex.Message}\n{ex.StackTrace}");
+                return false;
+            }
+            finally
+            {
+                // reset the callback for other http requests
+                ServicePointManager.ServerCertificateValidationCallback = null;
+            }
+        }
+
+        public bool CheckURL(string url, bool? useProxyAPI = null)
+        {
+            // if validation proxy enabled, access to the domain being validated is checked via our
+            // remote API rather than directly on the servers
+            bool useProxy = useProxyAPI ?? Certify.Properties.Settings.Default.EnableValidationProxyAPI;
+
+            //check http request to test path works
+            try
+            {
+                var request = WebRequest.Create(!useProxy ? url :
+                    Properties.Resources.APIBaseURI + "testurlaccess?url=" + url);
+                ServicePointManager.ServerCertificateValidationCallback = (obj, cert, chain, errors) =>
+                {
+                    // ignore all cert errors when validating URL response
+                    return true;
+                };
+                var response = (HttpWebResponse)request.GetResponse();
+
+                //if checking via proxy, examine result
+                if (useProxy)
+                {
+                    if ((int)response.StatusCode >= 200)
+                    {
+                        var encoding = ASCIIEncoding.UTF8;
+                        using (var reader = new System.IO.StreamReader(response.GetResponseStream(), encoding))
+                        {
+                            string jsonText = reader.ReadToEnd();
+                            Log("Proxy URL Check Result: " + jsonText);
+                            var result = Newtonsoft.Json.JsonConvert.DeserializeObject<Models.API.URLCheckResult>(jsonText);
+                            if (result.IsAccessible == true)
+                            {
+                                return true;
+                            }
+                        }
+                    }
+                    //request failed using proxy api, request again using local http
+                    return CheckURL(url, false);
+                }
+                else
+                {
+                    Log($"Local URL Check Result: HTTP {response.StatusCode}");
+                    //not checking via proxy, base result on status code
+                    return (int)response.StatusCode >= 200 && (int)response.StatusCode < 300;
+                }
+            }
+            catch (Exception)
+            {
+                System.Diagnostics.Debug.WriteLine("Failed to check url for access");
+                return false;
+            }
+            finally
+            {
+                // reset callback for other requests to validate using default behavior
+                ServicePointManager.ServerCertificateValidationCallback = null;
+            }
+        }
+    }
+}
